fix ninnyhammers
