+++ src/components/sessions/add-classes/AddClasses.tsx
@@ -1,25 +1,61 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-// eslint-disable-next-line import/order
-import React, { useState } from "react";
-import { Box, Typography } from "@material-ui/core";
+import React, { useState, useContext } from "react";
+
+import {
+  Box,
+  Button,
+  Divider,
+  IconButton,
+  MenuItem,
+  OutlinedInput,
+  Select,
+  Typography,
+} from "@material-ui/core";
+import { Add, RemoveCircle } from "@material-ui/icons";
 
 import { ClassListRequest } from "api/types";
+import FormRow from "components/common/form-row";
+import FieldVariant from "constants/FieldVariant";
+import QuestionType from "constants/QuestionType";
+import { UsersContext } from "context/UsersContext";
 
-import ClassForm, { ClassFormData } from "./ClassForm";
+import DaysPicker from "./DaysPicker";
+import useStyles from "./styles";
+
+// unique identifier for children form components
+let CLASS_COUNTER = 1;
+
+export const generateKey = (): number => {
+  let key = 0;
+  key = CLASS_COUNTER;
+  CLASS_COUNTER += 1;
+  return key;
+};
+
+export type ClassFormData = ClassListRequest & { index: number };
+
+export const defaultClassData: ClassListRequest = {
+  name: "",
+  days: [],
+  location: "",
+  facilitator: null,
+};
 
 type Props = {
-  classes: ClassListRequest[];
+  classList: ClassListRequest[];
   onChangeClasses: (classes: ClassListRequest[]) => void;
 };
 
-const AddClasses = ({ classes, onChangeClasses }: Props) => {
+const AddClasses = ({ classList, onChangeClasses }: Props) => {
+  const classes = useStyles();
+  const { users } = useContext(UsersContext);
+
   const classRequestToClassFormData = (
     classesData: ClassListRequest[]
   ): ClassFormData[] =>
     classesData.map((classData, i) => ({ ...classData, index: i }));
 
   const [classFormData, setClassFormData] = useState<ClassFormData[]>(
-    classRequestToClassFormData(classes)
+    classRequestToClassFormData(classList)
   );
 
   const handleChange = (classesData: ClassFormData[]) => {
@@ -29,6 +65,35 @@ const AddClasses = ({ classes, onChangeClasses }: Props) => {
       return req as ClassListRequest;
     });
     onChangeClasses(submitClassesData);
+  };
+
+  const onAddClass = (): void => {
+    handleChange([
+      ...classFormData,
+      { ...defaultClassData, index: generateKey() },
+    ]);
+  };
+
+  const onUpdateClass = (index: number, data: ClassListRequest): void => {
+    const classesObjs = [...classFormData];
+    classesObjs[index] = { ...classesObjs[index], ...data };
+    handleChange([...classesObjs]);
+  };
+
+  const onDeleteClass = (index: number): void => {
+    const classesObjs = classFormData;
+    classesObjs.forEach((classObj, i) => {
+      if (classObj.index === index) {
+        // Remove the class from array of classes
+        classesObjs.splice(i, 1);
+      }
+      if (classesObjs[i]) {
+        // Reset the index of all classes
+        classesObjs[i].index = i;
+      }
+    });
+    CLASS_COUNTER = classesObjs.length;
+    handleChange(classesObjs);
   };
 
   return (
@@ -42,11 +107,122 @@ const AddClasses = ({ classes, onChangeClasses }: Props) => {
         </Typography>
       </Box>
       <Box width={488}>
-        <ClassForm
-          dense={false}
-          onChange={(classesData) => handleChange(classesData)}
-          classesData={classFormData}
-        />
+        <Box display="flex">
+          <Box alignSelf="center" width="100%">
+            {classFormData.map((classData, i) => (
+              <Box key={classData.index} display="flex" marginBottom={1}>
+                <Box className={classes.rowContainer} minWidth={32}>
+                  {classFormData.length > 1 && (
+                    <IconButton
+                      aria-label="delete class"
+                      onClick={() => onDeleteClass(classData.index)}
+                      className={classes.deleteButton}
+                    >
+                      <RemoveCircle className={classes.deleteButton} />
+                    </IconButton>
+                  )}
+                </Box>
+                <Box flex="auto">
+                  <Box marginBottom={2}>
+                    <Typography variant="h4">
+                      Class {classData.index + 1}
+                    </Typography>
+                  </Box>
+                  <FormRow
+                    id={`classname-${classData.index}`}
+                    label="Name"
+                    questionType={QuestionType.TEXT}
+                    variant={FieldVariant.COMPACT}
+                  >
+                    <OutlinedInput
+                      autoComplete="new-password" // disable autocomplete
+                      className={classes.input}
+                      fullWidth
+                      id={`classname-${classData.index}`}
+                      placeholder="Class name"
+                      onChange={(e) =>
+                        onUpdateClass(i, { ...classData, name: e.target.value })
+                      }
+                      value={classData.name}
+                    />
+                  </FormRow>
+                  <DaysPicker
+                    days={classData.days}
+                    onChange={(days) => {
+                      onUpdateClass(i, { ...classData, days });
+                    }}
+                  />
+                  <FormRow
+                    id={`location-${classData.index}`}
+                    label="Location"
+                    questionType={QuestionType.TEXT}
+                    variant={FieldVariant.COMPACT}
+                  >
+                    <OutlinedInput
+                      autoComplete="new-password" // disable autocomplete
+                      className={classes.input}
+                      fullWidth
+                      id={`location-${classData.index}`}
+                      placeholder="Location"
+                      onChange={(e) =>
+                        onUpdateClass(i, {
+                          ...classData,
+                          location: e.target.value,
+                        })
+                      }
+                      value={classData.location}
+                    />
+                  </FormRow>
+                  <FormRow
+                    id={`class ${classData.index}`}
+                    label="Facilitator"
+                    questionType={QuestionType.SELECT}
+                    variant={FieldVariant.COMPACT}
+                  >
+                    <Select
+                      aria-label="Class facilitator"
+                      className={classes.input}
+                      displayEmpty
+                      fullWidth
+                      labelId={`class ${classData.index}`}
+                      onChange={(e) =>
+                        onUpdateClass(i, {
+                          ...classData,
+                          facilitator: e.target.value as number,
+                        })
+                      }
+                      value={classData.facilitator || ""}
+                      variant="outlined"
+                    >
+                      <MenuItem value="" className={classes.menuItem}>
+                        Facilitator
+                      </MenuItem>
+                      {users?.map((option) => (
+                        <MenuItem key={option.id} value={option.id}>
+                          {option.first_name} {option.last_name}
+                        </MenuItem>
+                      ))}
+                    </Select>
+                  </FormRow>
+                  {i < classFormData.length - 1 && (
+                    <Box paddingY={2}>
+                      <Divider />
+                    </Box>
+                  )}
+                </Box>
+              </Box>
+            ))}
+            <Button
+              onClick={onAddClass}
+              className={classes.addButton}
+              size="medium"
+              variant="outlined"
+            >
+              <Add fontSize="small" className={classes.addButtonIcon} />
+              Add class
+            </Button>
+          </Box>
+        </Box>
       </Box>
     </div>
   );
